payment And Razorpay

+++ SchoolManagementWebAPI/SchoolManagementWebAPI/Models/Class.cs
@@ -5,6 +5,6 @@
         public int Id { get; set; }
         public string ClassName { get; set; }
         public string TeacherCoordinatorId { get; set; }
-        public decimal Fees { get; set; }
+        public double Fees { get; set; }
     }
 }
